Airbrake configuration for 5.8.1
With the change to 5.8.1 our previous approach of only configuring
Airbrake if there is a ENV var seems to fail with "undefined method
`build_notice’ for nil:NilClass" errors - as it expects Airbrake to be
configured. This adds configuration in for all environments, but the
environment is ignored if there is not a ERRBIT_API_KEY env var.