Add non-technical explanations to responses
I've added a short explanation to each response.

The existing responses are good labels and are very useful as they are if you know what's going on.

So my approach has been to leave them in place and add a short gloss for people who don't know what's going on, for example:

CHANGED "Unsupported scheme"
TO "Unsupported scheme - we can't check this type of link."

or

CHANGED  "Not a valid URI."
TO  "Not a valid URI or URL."
(No one seems to understand 'URI').

I've tried to be clever on lines 97 and 134 and use string interpolation. Shout if it's wrong/a bad idea.

Redirects could probably do with some more work.