Implement HistoricalLinkChecker
`HistoricalLinkChecker` wraps `LinkChecker` to look back on previous check results so that we can raise a different set of errors when a particular error has been recurring for a while.

This will enable us to escalate and prioritise issues that have occurred for a longer period of time.

For example, in future, we can prioritise 500 or 404 errors that have been occurring for more than a week over ones that have only been happening for a day. Sometimes URLs will disappear temporarily when a website is being deployed, or error temporarily when a developer deploys a bug (whoops).