Pass the Faraday http client around during a link test
This is to fix a thread safety issue where an ||= is run at class level
as ||= is not thread safe.

It seems unnecessary to require this at class level as we only care
about cookie jar for the duration of a links check. This therefore
introduces the concept of a http client at the checker level which can
be passed around to an additional instances of checker that are created.