Ensure WebhookWorker jobs are unique
We recently downgraded sidekiq-unique-jobs to use our own fork,
but in the process, we added a 'require: false' to the Gemfile.
This meant that sidekiq-unique-jobs wasn't being loaded and
therefore not being used. This has lead to the queue regularly
hitting 2 million jobs at night and increasing the memory usage
of Redis significantly.