Update batch tests to check authorisation
Having a user created before every test meant that we can’t test for unauthorised users
trying to create batch reports.

Batch request tests have been updated to pass in a bearer token as that is how applications
will be authenticated in the “real world”

Removed the authentication helper as it’s not needed. There aren’t any controller tests that need to
bypass user authentication, and the request tests allow you to pass a header into the post request.