Treat SSL issues as warnings
There are many local authority websites which have misconfigured SSL
certs: they only send the leaf cert and not the full chain.  This is
incorrect according to the standard, and SSL checker tools (like
ssllabs.com) will flag this up.  But browsers will often work with
such websites: they cache certs, and so if the user has visited a
website which does serve the intermediate cert, then the website which
doesn't will work.

Since a user visiting the page in a browser gets the page, the link
"works", even though it technically has a problem.  There are a lot of
websites with this issue, so it really diminishes the value of the
link-checker-api to report these as hard errors.

We're not in the business of monitoring other people's websites for
configuration issues, so we're downgrading this issue to a warning.