Use normalised URI when making HTTP requests
When making an HTTP request for a URI we should make sure to normalise
it first otherwise the built in Ruby Net::HTTP module isn't able to
perform the request.

This is similar to what browsers would do anyway when receiving an
invalid but normalisable URI, so it should help to reduce the number of
invalid URIs we get which would actually work.