Disable CSRF protection
In Rails 5.2, CSRF protection was turned on by default. See this commit [1]

Running Rails app:update has updated all the defaults to match 6.1.4, including
turning on the defaults for the previous Rails versions. So whilst
config.action_controller.default_protect_from_forgery hasn't been changed in the latest
release, it's inadvertantly been turned on here.

We don't need CSRF protection, as this is an API so it's normal that it is responding
to cross site requests from other applications. Nice explanation [2]

The security strategy employed by link checker api and its consumers is using a :webhook_secret_token [3]

1 - https://github.com/rails/rails/commit/ec4a836919c021c0a5cf9ebeebb4db5e02104a55
2 - https://stackoverflow.com/questions/35181340/rails-cant-verify-csrf-token-authenticity-when-making-a-post-request
3 - https://github.com/alphagov/link-checker-api/blob/main/docs/api.md#verifying-the-webhook-request