Clean database after Pact tests
This commit changes the configuration of the Pact tests.

They were previously configured to truncate the database before each Pact example. However, this meant that the final example in the Pact test suite wouldn't clean up after itself, causing records to be left in the database.

This resulted in subsequent test runs to fail because the test suite expected to begin with an empty database.

This change configures Pact to use the 'transaction' strategy of Database Cleaner to run each example in its own database transaction. The database transaction is started before each Pact test, and is rolled back after each test finishes.

The result is that the database remains clean after a full run of the test suite.