Disable sidekiq-unique-jobs in tests
This was previously causing flakey builds, where state from one test was affecting later tests. It is also following a recommendation in the gem not to test the unique functionality (see "https://github.com/mhenrixon/sidekiq-unique-jobs#uniqueness"). This commit removes a test for the webhook worker which was contributing to this flakiness, and instead adds tests that ensure the sidekiq options provided are valid.