Do not raise when a webhook experiences a remote server error
This application returns the link checking results to the requesting
application through a webhook.

In some cases, the remote server will respond with an error (e.g. a
timeout or some other 5xx status code). We don't need to log these to
Sentry, since they are outside of this application's control.

Therefore rescuing from the `Faraday::ServerError` raise that can occur
in this worker.