Fix faulty shared test example
In the `has errors` shared test example, we were passing in a string to
check against the errors but not actually checking it.

This meant you could add any string and the test would always pass,
provided an error existed.

Therefore updating the shared example to actually accept a parameter,
check the error is returned and update the messages to reflect what the
class is actually returning to users.