Fix Rails secrets deprecation warning
Rails 7.1 deprecates the use of Rails.application.secrets in favour of
Rails.application.credentials.

The credentials system introduces the burden of master encryption key
administration at no benefit to us, because we manage our production
secrets using environment variables instead of committing them to Git.

This commit loads the existing secret values and merges them into the
credentials object. This approach was copied from
https://github.com/alphagov/account-api/commit/8937b172be530a5d91bd999f4538e5a722dcab19.

Performing this change now, as we never did this to remove the
deprecation warnings when upgrading to Rails 7.1.