Report pointless redirects
We currently report to users when a link exceeds 2 redirects, however we
are finding users adding links that include one redirect.

This is pointless as we were finding users putting links to Outlook
redirects in content.

Therefore reporting on these pointless redirects, so users can just use
the destination URL instead.