Add new 'guid' column to the Batches table
Using a sequential `batch_id` has been a real headache. It led to
code being added to Whitehall to handle errors that happen in
Staging and Integration after a data sync, because the Whitehall
and Link Checker API databases would not copy over at the same
time:
https://github.com/alphagov/whitehall/pull/3675

I was able to remove that code by flipping the order of the data
sync, so that Whitehall is always 'behind' Link Checker API, and
therefore never receives a batch_id that it has encountered
already (and thus hitting a uniqueness constraint issue):
https://github.com/alphagov/govuk-helm-charts/pull/3058

But in local testing, we're still having issues, because we're
having to have a fully populated Link Checker API database in
order to not run into the same uniqueness clash. And the data
replication script in govuk-docker isn't working as the restore
was done from a different version of Postgres than is running
locally.

Rather than continue to try to work around the issue, a better
architecture would be to generate a truly unique batch_id every
time. That way, we never have to worry about uniqueness or
sequencing.

So we're adding a new guid field, and will make changes to
Whitehall to refer to this instead of to its batch_id.