Don't make HTTP requests to suspicious domains
The Link Checker API logic was continuing to go through all the
other link checks (checking for redirects, reachability, etc).
But suspicious domains can be dangerous domains. We don't want to
send a HTTP request to those domains, period.

So now we return early if a link is hosted on a suspicious domain.

Trello: https://trello.com/c/tmnht4P1