Fetch suspicious domains from DB instead of config
We're going to add a bunch more suspicious domains to our checks.
It's not necessarily a good idea to make that list public.
Therefore we're moving the list of domains into a database (and
benefiting from increased validation to boot 👢).

Post merge, we'll run this in the Rails console so that we have
parity with current behaviour:

```
SuspiciousDomain.create([
  { domain: "nostringsng.com" },
  { domain: "becauseiamagirl.org" },
  { domain: "bilebrizoua.ci" }
])
```

Trello: https://trello.com/c/tmnht4P1