Memoize old checks in `CleanupJob`
We were querying for checks to delete twice, once using the result to
delete foreign key relations and the other time to delete the checks
themselves.

It is possible to checks meeting the criteria changed between both runs,
resulting in a foreign key constraint error when deleting some checks as
the batch had not been deleted.

Memoizing the old checks query result, so we always target the same
records on both deletions.