Revert `RestartJobException` to be excluded from Sentry
In b74c63744d63afc32d3d6498f8cf0e92fa9c4033, the
`RestartWorkerException` (now `RestartJobException`) was moved from
`excluded_exceptions` to `data_sync_excluded_exceptions`. This had the
effect of meaning these exceptions (which are solely client error, that
are unactionable by us) were reported to Sentry when they occurred
outside the data sync.

Therefore adding this exception back into `excluded_exceptions` so it is
not reported at any time of day. It is also being left in
`data_sync_excluded_exceptions` as the commit that moved it suggested
this was being logged during the data sync.